+++ src/core/interpreter/mod.rs
@@ -1,0 +1,6 @@
+pub mod arithmetic;
+pub mod blocks;
+pub mod boolean;
+pub mod functions;
+pub mod interpreter;
+pub mod misc_proc;
